misc(source): rename files to Deno convention
