+++ .gitignore
@@ -101,3 +101,4 @@ web_modules/
 *.iml
 
 npm/
+swagger/
