+++ deno.jsonc
@@ -2,7 +2,7 @@
 	"tasks": {
 		"build": "deno task generate:swagger && deno task test && deno task generate:npm-package",
 		"generate:npm-package": "deno run -A scripts/build_npm.ts",
-		"generate:swagger": "npx -y swagger-typescript-api -p https://swagger.getlago.com/swagger.yaml -o ./src/swagger -n client.ts",
-		"test": "deno test src/test --parallel"
+		"generate:swagger": "npx -y swagger-typescript-api -p https://swagger.getlago.com/swagger.yaml -o ./swagger -n client.ts",
+		"test": "deno test ./tests --parallel"
 	}
 }
