+++ tests/wallet.test.ts
@@ -3,7 +3,7 @@ import type {
   WalletInput,
   Wallets,
   WalletUpdateInput,
-} from "../main.ts";
+} from "../mod.ts";
 import { lagoTest, unprocessableErrorResponse } from "./utils.ts";
 
 const walletInput = {
